Moved DTOs from single file to separate files per DTO for Citizens & Institutions
