+++ backend/FlashIdBackend/Application/Features/Citizens/DTOs/RegisterCitizenRequestDto.cs
@@ -7,12 +7,3 @@ public class RegisterCitizenRequestDto
     public string Password { get; set; } = string.Empty;
     public string ActivationCode { get; set; } = string.Empty;
 }
-
-public class RegisterCitizenResponseDto
-{
-    public Guid CitizenId { get; set; }
-    public Guid UserId { get; set; }
-    public string SaId { get; set; } = string.Empty;
-    public string Username { get; set; } = string.Empty;
-    public DateTime CreatedAt { get; set; }
-}
