+++ backend/FlashIdBackend/Application/Features/Institutions/DTOs/RegisterInstitutionRequestDto.cs
@@ -1,0 +1,11 @@
+using Domain.Enums;
+
+namespace Application.Features.Institutions.DTOs;
+
+public class RegisterInstitutionRequestDto
+{
+    public string Name { get; set; } = string.Empty;
+    public InstitutionType Type { get; set; }
+    public string VerificationNumber { get; set; } = string.Empty;
+    public Guid AdminId { get; set; }
+}
